# MindApi/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "This is the Master Inventory Network Database (MIND) API. MIND provides automated allocation of technical attributes needed to create IP and Ethernet VPNs on the global Sigma network. MIND supports the 'Nova' services specfication which defines the collection of connectivity services supported by ENT. Go to https://thehub.thomsonreuters.com/docs/DOC-2193014 to learn more.",
    "version" : "1.0.0",
    "title" : "MIND API",
    "contact" : {
      "email" : "jonathan.beasley@thomsonreuters.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "raid.thomsonreuters.com",
  "basePath" : "/v1",
...
  "schemes" : [ "https", "http" ],
  "paths" : {
...
    "/infrastructure/attachment/{attachmentId}/vif" : {
      "get" : {
        "tags" : [ "infrastructure vif" ],
        "summary" : "Find all vifs for a given attachment",
        "description" : "Returns all vifs for a given attachment",
        "operationId" : "getAllInfrastructureVifsByAttachmentId",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "attachmentId",
          "in" : "path",
          "description" : "ID of the attachment",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Vif"
              }
            }
          },
          "404" : {
            "description" : "The specified resource was not found",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "infrastructure vif" ],
        "summary" : "Create a new vif",
        "description" : "",
        "operationId" : "addInfrastructureVif",
        "consumes" : [ "application/json", "application/xml" ],
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "attachmentId",
          "in" : "path",
          "description" : "ID of the attachment",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "vif request object that generates a new vif",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/InfrastructureVifRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "schema" : {
              "$ref" : "#/definitions/Vif"
            }
          },
          "400" : {
            "description" : "Validation error",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          },
          "404" : {
            "description" : "The specified resource was not found",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/infrastructure/attachment/vif/{vifId}" : {
      "get" : {
        "tags" : [ "infrastructure vif" ],
        "summary" : "Find vif by ID",
        "description" : "Returns a single vif",
        "operationId" : "getInfrastructureVifById",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "vifId",
          "in" : "path",
          "description" : "ID of the vif",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "schema" : {
              "$ref" : "#/definitions/Vif"
            }
          },
          "404" : {
            "description" : "The specified resource was not found",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "infrastructure vif" ],
        "summary" : "Update an existing vif",
        "description" : "",
        "operationId" : "updateInfrastructureVif",
        "consumes" : [ "application/json", "application/xml" ],
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "vifId",
          "in" : "path",
          "description" : "ID of the vif",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "vif update object that updates an existing vif",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/VifUpdate"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "schema" : {
              "$ref" : "#/definitions/Vif"
            }
          },
          "400" : {
            "description" : "Validation error",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          },
          "404" : {
            "description" : "The specified resource was not found",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "infrastructure vif" ],
        "summary" : "Deletes a vif",
        "description" : "",
        "operationId" : "deleteInfrastructureVif",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "vifId",
          "in" : "path",
          "description" : "ID of the vif",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "204" : {
            "description" : "Successful operation",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          },
          "404" : {
            "description" : "The specified resource was not found",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
    "ApiResponse" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "type" : {
          "type" : "string"
        },
        "message" : {
          "type" : "string"
        }
      },
      "example" : {
        "code" : 0,
        "type" : "type",
        "message" : "message"
      }
    },
...
    "InfrastructureVifRequest" : {
      "type" : "object",
      "properties" : {
        "isLayer3" : {
          "type" : "boolean",
          "description" : "Determines if the vif should be enabled for layer 3",
          "default" : false
        },
        "vlanTag" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "IEEE 802.1Q vlan tag allocated to the vif",
          "minimum" : 1,
          "maximum" : 4094
        },
        "IpAddress1" : {
          "type" : "string",
          "description" : "IPv4 address assigned to the first connection in the vif",
          "pattern" : "^(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$"
        },
        "SubnetMask1" : {
          "type" : "string",
          "description" : "IPv4 subnet mask assigned to the first connection in the vif",
          "pattern" : "^(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$"
        },
        "IpAddress2" : {
          "type" : "string",
          "description" : "IPv4 address assigned to the second connection in the vif",
          "pattern" : "^(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$"
        },
        "SubnetMask2" : {
          "type" : "string",
          "description" : "IPv4 subnet mask assigned to the second connection in the vif",
          "pattern" : "^(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$"
        },
        "IpAddress3" : {
          "type" : "string",
          "description" : "IPv4 address assigned to the third connection in the vif",
          "pattern" : "^(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$"
        },
        "SubnetMask3" : {
          "type" : "string",
          "description" : "IPv4 subnet mask assigned to the third connection in the vif",
          "pattern" : "^(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$"
        },
        "IpAddress4" : {
          "type" : "string",
          "description" : "IPv4 address assigned to the fourth connection in the vif",
          "pattern" : "^(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$"
        },
        "SubnetMask4" : {
          "type" : "string",
          "description" : "IPv4 subnet mask assigned to the fourth connection in the vif",
          "pattern" : "^(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$"
        }
      },
      "example" : {
        "vlanTag" : 328,
        "SubnetMask1" : "SubnetMask1",
        "SubnetMask2" : "SubnetMask2",
        "SubnetMask3" : "SubnetMask3",
        "SubnetMask4" : "SubnetMask4",
        "isLayer3" : false,
        "IpAddress1" : "IpAddress1",
        "IpAddress2" : "IpAddress2",
        "IpAddress3" : "IpAddress3",
        "IpAddress4" : "IpAddress4"
      },
      "xml" : {
        "name" : "InfrastructureVifRequest"
      }
    },
    "Vif" : {
      "type" : "object",
      "properties" : {
        "vifId" : {
          "type" : "integer",
          "description" : "The ID of the vif"
        },
        "name" : {
          "type" : "string",
          "description" : "The name of the vif  "
        },
        "isLayer3" : {
          "type" : "boolean",
          "description" : "Determines if the vif is enabled for layer 3"
        },
        "vlanTag" : {
          "type" : "integer",
          "description" : "The vlan tag number assigned to the vif"
        },
        "attachmentId" : {
          "type" : "integer",
          "description" : "The ID of the parent attachment"
        },
        "tenantId" : {
          "type" : "integer",
          "description" : "The ID of the owning tenant"
        },
        "routingInstance" : {
          "$ref" : "#/definitions/RoutingInstance"
        },
        "vlans" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Vlan"
          }
        },
        "contractBandwidthPool" : {
          "$ref" : "#/definitions/ContractBandwidthPool"
        }
      },
      "example" : {
        "vlans" : [ {
          "vlanTag" : 2,
          "vlanID" : 5
        }, {
          "vlanTag" : 2,
          "vlanID" : 5
        } ],
        "vlanTag" : 6,
        "contractBandwidthPool" : {
          "name" : "name",
          "contractBandwidthMbps" : 5
        },
        "name" : "name",
        "tenantId" : 5,
        "isLayer3" : true,
        "attachmentId" : 1,
        "routingInstance" : {
          "routingInstanceId" : 0,
          "name" : "name"
        },
        "vifId" : 0
      },
      "xml" : {
        "name" : "Vif"
      }
    },
    "VifUpdate" : {
      "type" : "object",
      "properties" : {
        "contractBandwidthPoolName" : {
          "type" : "string",
          "description" : "The name of an existing contract bandwidth pool. Changing the contract bandwidth pool will cause this vif to share bandwidth with other vifs in the same contract bandwidth pool",
          "maxLength" : 32
        },
        "contractBandwidthMbps" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "Select a contract bandwidth if you want this vif to be allocted its own dedicated contract bandwidth pool. If a contract bandwidth is selected, the provider will create a new contract bandwidth pool for this vif"
        },
        "routingInstanceName" : {
          "type" : "string",
          "description" : "The name of an existing routing instance which will be used by the vif for the routing IP packets",
          "maxLength" : 32
        },
        "createNewRoutingInstance" : {
          "type" : "boolean",
          "description" : "Determines if the provider should create a new routing instance for the vif. This is useful if the vif currently shares a routing instance with one or more other vifs and it is desired to give this vif its own routing instance",
          "default" : false
        },
        "trustReceivedCosDscp" : {
          "type" : "boolean",
          "description" : "Determines whether DSCP and COS markings of packets should be trusted by the provider",
          "default" : false
        }
      },
      "example" : {
        "createNewRoutingInstance" : false,
        "routingInstanceName" : "routingInstanceName",
        "contractBandwidthMbps" : 0,
        "trustReceivedCosDscp" : false,
        "contractBandwidthPoolName" : "contractBandwidthPoolName"
      },
      "xml" : {
        "name" : "VifUpdate"
      }
    },
    "RoutingInstance" : {
      "type" : "object",
      "properties" : {
        "routingInstanceId" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "The ID of the routing instance"
        },
        "name" : {
          "type" : "string",
          "description" : "The name of the routing instance"
        }
      },
      "example" : {
        "routingInstanceId" : 0,
        "name" : "name"
      },
      "xml" : {
        "name" : "RoutingInstance"
      }
    },
    "ContractBandwidthPool" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string",
          "description" : "The name of the contract bandwidth pool "
        },
        "contractBandwidthMbps" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "The contract bandwidth of the pool in Mbps"
        }
      },
      "example" : {
        "name" : "name",
        "contractBandwidthMbps" : 5
      },
      "xml" : {
        "name" : "ContractBandwidthPool"
      }
    },
    "Vlan" : {
      "type" : "object",
      "properties" : {
        "vlanID" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "The ID of the vlan"
        },
        "vlanTag" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "The vlan tag"
        }
      },
      "example" : {
        "vlanTag" : 2,
        "vlanID" : 5
      }
    },
...
  },
...
}